+++ NEWS.md
@@ -1,5 +1,6 @@
-# radiant.model 0.9.7.7
+# radiant.model 0.9.7.8
 
+* Ensure variable and dataset names are valid for R (i.e., no spaces or symbols), "fixing" the input as needed
 * Cross validation functions for decision trees (`cv.crtree`) and neural networks(`cv.nn`) that can use various performance metrics for during evaluation e.g., `auc` or `profit`
 * Option to add square and cube terms in _Model > Linear regression_ and _Model > Logistic regression_.
 * Option to pass additional arguments to `shiny::runApp` when starting radiant such as the port to use. For example, radiant.model::radiant.model("https://github.com/radiant-rstats/docs/raw/gh-pages/examples/demo-dvd-rnd.state.rda", port = 8080)